Add support for background color of inactive keyboard focus list items
When navigating the file explorer list with the keyboard and moving the
focus to another UI component (like the editor) the background color of
the selected item is colorized with the `list.inactiveFocusBackground`
theme key.
It has been set to `nord2` with a opacity of 80% to match the style of
inactive list items that were focused without the keyboard
(`list.inactiveSelectionBackground`).

GH-107