OSS community standards: Code of Conduct & Contributing Guidelines
The contribution guidelines (1) help to build a community that
encourages people to use, contribute to (2), and evangelize a project.

It includes sections about

* how to get started
* bug reports
* enhancement suggestions
* pull requests
* style guides
  *  JavaScript Style Guide (3)
  *  Markdown Style Guide (4)
  * Git Style Guide (5)
* credits

See the GitHub introduction blog post (1) and GitHub Help (6) for more
details about the provided integrations.

To facilitate a healthy and constructive community behavior, Nord Visual
Studio Code adheres and enforce a code of conduct (7).

This commit includes the sections about

* what we believe in and how we act
* unacceptable behavior
* the responsibilities of the maintainer
* the enforcement of the Code of Conduct
* consequences for violations
* the scope of the Code of Conduct

See the GitHub Help (8) for more details about the provided
integrations.

References:
  (1) https://github.com/blog/1184-contributing-guidelines
  (2) https://opensource.guide/how-to-contribute
  (3) https://github.com/arcticicestudio/styleguide-javascript
  (4) https://github.com/arcticicestudio/styleguide-markdown
  (5) https://github.com/arcticicestudio/styleguide-git
  (6) https://help.github.com/articles/setting-guidelines-for-repository-contributors
  (7) https://opensource.guide/code-of-conduct
  (8) https://help.github.com/articles/adding-a-code-of-conduct-to-your-project

Resolves GH-131