Rename theme file to enable theme development features
An undocumented feature for theme extension developers is to ensure the
name of the JSON file is suffixed with `-color-theme.json`.
This enables the JSON scheme validation for the theme API allowing
developers to validate the implemented theme keys, showing warnings
about deprecated keys and providing full auto completion,
field documentations and color previews (color picker) for the HEX
format.
This change aligns Nord to the official bundled and default themes (1)
by adapting to the naming scheme without being a breaking change since
the theme is identified by it's extension ID as well as the
`_metadata` field in the `package.json` and not by the name of the theme
file (which would be odd since a theme can provide multiple theme
files).

References:
  (1) https://github.com/microsoft/vscode/tree/master/extensions/theme-abyss/themes
  (2) https://github.com/svipas

Extracted from GH-143 submitted by @svipas (2)
Resolves GH-148