Filled background color for "Find" widget's button toggle active… (#153)
In VS Code 1.37 (1) (July 2019) the Find widget's button toggle active
state now has a filled background so that it is easier to tell when the
focus is on an active toggle.

To customize the background color of the toggle active state, the new
`inputOption.activeBackground` UI/workbench theme key has been added to
Nord.

Unfortunately there is no UI theme key to also change the foreground
color of active state toggles. It is "hard-coded" using a relatively
bright color so it is not possible to also use a bright background
color. It would be great to apply a "reverse" effect to e.g. use
`nord0` as foreground to increase the contrast when using `nord8`
as background color.

References:
  (1) https://code.visualstudio.com/updates/v1_37#_button-toggle-active-state-in-find-widget

Resolves GH-151