Add text color related workbench theme keys
Added the `descriptionForeground` workbench theme key that was
introduced almost 2 years ago in microsoft/vscode@f450e0b1fe3 (1) for
microsoft/vscode#26298 (2), but not mentioned in a changelog so there
was no support for it in Nord.
It now uses `nord4` with a opacity of 90% as color value (`#d8dee9e6`).

References:
  (1) https://github.com/microsoft/vscode/blame/f450e0b1fe3ac26ce6bf8367ff1bd60fb1387a7d/src/vs/platform/theme/common/colorRegistry.ts#L154
  (2) https://github.com/microsoft/vscode/issues/26298

GH-160