Adopt semantic highlighting
Visual Studio Code version 1.44 introduced support [1] for
semantic highlighting [2] that allows to assign colors and
styles to tokens. Semantic highlighting enriches syntax coloring based
on symbol information from the language service, which has more
complete understanding of the project so the coloring changes appear
once the language server is running and has computed the semantic tokens.

The feature has been tested for some weeks and worked out-of-the-box
without the need to change or add specific matchers or rules.

[1]: https://code.visualstudio.com/updates/v1_44#_semantic-tokens-provider-api
[2]: https://code.visualstudio.com/docs/getstarted/themes#_semantic-highlighting
[3]: https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide#theming

GH-172