Activity Bar indicator styles
VS Code 1.40 [1] (October 2019) introduced a new indicator for the
active item in the Activity Bar [2] to make it stand out better and
increase readability. This includes an optional background color for
the active element.

The new theme keys `activityBar.activeBorder` and
`activityBar.activeBackground` have been added to make it conform with
Nord's theme style.

[1]: https://code.visualstudio.com/updates/v1_22
[2]: https://code.visualstudio.com/updates/v1_40#_activity-bar-indicator

GH-174

Co-authored-by: Sven Greb <development@svengreb.de>