Color token for debugged line shown in overview ruler
VS Code 1.52 [1] (November 2020) introduced the new
`editor.focusedStackFrameHighlightBackground` theme color token [2] to
highlight the currently focused debugged line in the overview ruler.
Additionally, the `editor.stackFrameHighlightBackground` key has been
added that is used to highlight the decoration in the overview ruler on
the right of the editor.
Both tokens using `nord10` to make them more conform with Nord's style.

[1]: https://code.visualstudio.com/updates/v1_52
[2]: https://code.visualstudio.com/updates/v1_52#_ui-improvements

Co-authored-by: Sven Greb <development@svengreb.de>

GH-205