Color tokens for bracket pair colorization
VS Code 1.60 [1] (August 2021) introduced native, high performance
bracket pair colorization [2] (opt-in via
`editor.bracketPairColorization.enabled` setting) to address performance
issues of popular extensions.
The color tokens have been added and adjusted to match Nord's theme
style:

- `statusBarItem.warningBackground` - the background color used for
  warning items in the status bar that stand out from other status bar
  entries to indicate warning conditions.
- `statusBarItem.warningForeground` - the foreground color used for
  warning items in the status bar that stand out from other status bar
  entries to indicate warning conditions.

Note that bracket pair colorization is a feature that does not go well
with the design concepts of Nord and the recommendation is therefore to
not enable it. Of course this is always a personal preference and the
feature be used as you wish, but note that the bracket colors defined by
the Nord theme won't be changed on request but users can simply
customize it on their own [3].

[1]: https://code.visualstudio.com/updates/v1_60
[2]: https://code.visualstudio.com/updates/v1_60#_high-performance-bracket-pair-colorization
[3]: https://www.nordtheme.com/docs/ports/visual-studio-code/customization

Co-authored-by: Sven Greb <development@svengreb.de>

GH-225